Remove unnecessary usage of app_ctx_stack
This was only used to check if the application context exists, and if it didn't
exist an exception was raised with an explanatory message.

Trying to access current_app._get_current_object() raises the same kind of
exception with an equivalent explanatory message, therefore the usage of
app_ctx_stack and the check around it was completely redundant.